feat(avtool): updates for clarity
