+++ experiments/mcp-genmedia/mcp-genmedia-go/mcp-avtool-go/gcs_utils.go
@@ -1,0 +1,116 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"log"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"cloud.google.com/go/storage"
+)
+
+func downloadFromGCS(ctx context.Context, gcsURI, localDestPath string) error {
+	bucketName, objectName, err := parseGCSPath(gcsURI)
+	if err != nil {
+		return err
+	}
+
+	client, err := storage.NewClient(ctx)
+	if err != nil {
+		return fmt.Errorf("storage.NewClient: %w", err)
+	}
+	defer client.Close()
+
+	gcsOpCtx, cancel := context.WithTimeout(ctx, 2*time.Minute)
+	defer cancel()
+	rc, err := client.Bucket(bucketName).Object(objectName).NewReader(gcsOpCtx)
+	if err != nil {
+		return fmt.Errorf("Object(%q).NewReader: %w", objectName, err)
+	}
+	defer rc.Close()
+
+	destDir := filepath.Dir(localDestPath)
+	if err := os.MkdirAll(destDir, 0755); err != nil {
+		return fmt.Errorf("os.MkdirAll for directory %s: %w", destDir, err)
+	}
+
+	f, err := os.Create(localDestPath)
+	if err != nil {
+		return fmt.Errorf("os.Create: %w", err)
+	}
+	defer f.Close()
+
+	if _, err := io.Copy(f, rc); err != nil {
+		return fmt.Errorf("io.Copy: %w", err)
+	}
+	log.Printf("Successfully downloaded %s to %s", gcsURI, localDestPath)
+	return nil
+}
+
+func uploadToGCS(ctx context.Context, bucketName, objectName, contentType string, data []byte) error {
+	client, err := storage.NewClient(ctx)
+	if err != nil {
+		return fmt.Errorf("storage.NewClient: %w", err)
+	}
+	defer client.Close()
+
+	obj := client.Bucket(bucketName).Object(objectName)
+	wc := obj.NewWriter(ctx)
+
+	finalContentType := contentType
+	if finalContentType == "" {
+		ext := strings.ToLower(filepath.Ext(objectName))
+		switch ext {
+		case ".mp3":
+			finalContentType = "audio/mpeg"
+		case ".wav":
+			finalContentType = "audio/wav"
+		case ".mp4":
+			finalContentType = "video/mp4"
+		case ".mov":
+			finalContentType = "video/quicktime"
+		case ".mkv":
+			finalContentType = "video/x-matroska"
+		case ".webm":
+			finalContentType = "video/webm"
+		case ".png":
+			finalContentType = "image/png"
+		case ".jpg", ".jpeg":
+			finalContentType = "image/jpeg"
+		case ".gif":
+			finalContentType = "image/gif"
+		default:
+			log.Printf("uploadToGCS: Could not infer ContentType for extension '%s' of object '%s'. Uploading without explicit ContentType.", ext, objectName)
+		}
+	}
+
+	if finalContentType != "" {
+		wc.ContentType = finalContentType
+		log.Printf("uploadToGCS: Setting ContentType to '%s' for object '%s'", finalContentType, objectName)
+	}
+
+	if _, err := wc.Write(data); err != nil {
+		wc.Close()
+		return fmt.Errorf("Writer.Write: %w", err)
+	}
+	if err := wc.Close(); err != nil {
+		return fmt.Errorf("Writer.Close: %w", err)
+	}
+	return nil
+}
+
+func parseGCSPath(gcsURI string) (bucketName, objectName string, err error) {
+	if !strings.HasPrefix(gcsURI, "gs://") {
+		return "", "", fmt.Errorf("invalid GCS URI: must start with 'gs://', got %s", gcsURI)
+	}
+	trimmedURI := strings.TrimPrefix(gcsURI, "gs://")
+	parts := strings.SplitN(trimmedURI, "/", 2)
+	if len(parts) < 2 || parts[0] == "" || parts[1] == "" {
+		return "", "", fmt.Errorf("invalid GCS URI format: %s. Expected gs://bucket/object", gcsURI)
+	}
+	return parts[0], parts[1], nil
+}
